requirements.txt
python-dotenv
psycopg2
requests
beautifulsoup4
python-telegram-bot